Try to cross build for Scala3
The issue is that the fs2-cron-cron4s subproject is not working for
Scala 3, because cron4s is not available for scala 3 (and it is
discouraged to provide libraries for Scala3 that have Scala 2
dependencies). So the goal is to configure sbt such that it builds
core and calev subprojects for all scala versions and the cron4s
subproject only for 2.12 and 2.13.